requirements.txt
# common
mysqlclient==2.1.1
django==4.1.1
django-redis==5.2.0

# djangorestframework
djangorestframework==3.13.1
markdown==3.4.1
django-filter==22.1

# swagger
drf-yasg2==1.19.4